backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
PyJWT==2.10.1
redis==5.2.1
